Variables needed for implementation of the model equations added.
Mechanical coefficients, logical variables to identify derived classes
have been added in most of the classes.
However the model equations cannot be implemented yet. Triangle and
Centre class require a pointer to a Cell object, and when added they
mess up all the dependencies. Now I need to fix the dependency tree
before continuing with the implementation of the model.